fix(envelope): created_at must be int, not float — matches canonical hash
Real bug uncovered while running gate-server tests. EnvelopeOut.created_at
was typed as float (default 0.0) but maelstrom-gate's canonical
AuthorizationEnvelope dataclass uses int (Unix microseconds from
time_ns() // 1000).

Pydantic preserved the float type on the round-trip:
  build:  created_at = 1777498795572774         (int)
  json:   "created_at": 1777498795572774        (no decimal)
  verify: pydantic decodes float → 1777498795572774.0
  json:   "created_at": 1777498795572774.0      (".0" appended!)

The trailing ".0" produced different canonical-hash JSON than the
build path, so HMAC verification ALWAYS failed for the built envelope.
Two tests caught it:
  test_envelope_verify_valid          (the round-trip case)
  test_register_filter_validate_envelope_cycle  (full workflow)

Fix: change EnvelopeOut.created_at type from float → int. Pydantic now
keeps it as int through the round-trip; canonical hash matches; signature
verifies. 34/34 gate-server tests pass.

This was production-breaking — every built envelope would fail
verification, so the entire envelope-based authorization flow was
non-functional.

+++ gate_server/models.py
@@ -69,7 +69,9 @@ class EnvelopeOut(BaseModel):
     dry_run: bool
     branching: str
     human_approved: bool
-    created_at: float = 0.0
+    created_at: int = 0  # Unix microseconds — must be int to match canonical
+                         # AuthorizationEnvelope serialization. A float here
+                         # adds ".0" to JSON, breaking signature reverification.
     signature: str
